requirements.txt
asgiref==3.3.4
autopep8==1.5.7
boto3==1.18.0
botocore==1.21.0
Django==3.2.4
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6